Clear window title after executing command
- This is required if there are multiple events trigerring same command.
For example, in applet page, there are multiple applets with reference
to budgie-extra repository. If we click on any of them, and click on
other with the same link, it won't work. Because, both commands are same
and there is no effective title change. Clearing title after each
command execution can fix this issue.